NN_python/requirements.txt
pytorch_lightning
torch
numpy
opencv-python
tifffile
torchvision
tqdm
tensorboardX
tensorboard
cuda-python
bresenham
opencv-python
matplotlib
Pillow
